feat(up)!: add `--upgrade` parameter (#723)
`omni up` now takes a new `--upgrade` parameter which forces checking
new available versions of tools matching any requested tool and version
constraint, and installing them if it is not already the case.

This affects any tool which can be upgraded, e.g. this affects the
`homebrew` operation, the `github release` operation, and any
`asdf`-backed operation. For each of those, a new `upgrade` boolean
parameter has been added and can be set to `true` to force the
`--upgrade` behavior in all cases. When using `latest` as the version,
this will only match with any version of the latest `major`.

A new `up_command`/`upgrade` option is also available in the global/repo
configuration and can be set to `true` to force upgrading all tools even
without the `--upgrade` parameter.

BREAKING CHANGE: This changes the default behavior which would now be
the equivalent of setting `up_command`/`upgrade` to `true`. The default
behavior was changed to speed up the operations.

Closes https://github.com/XaF/omni/issues/705