feat(argparser): ✨ omni can now parse arguments for custom commands
Until now, when writing a custom command for omni, one had to:
- use the language-specific argument parser
- report the supported arguments in the omni metadata

This led to:
- discrepancies in UX if writing commands in multiple languages
- writing how arguments are parsed at least twice (arg parser + omni
  metadata)

This introduces a new `argparser` metadata for custom commands, as well
as a number of new configuration for the `parameters` declared in the
metadata. When `argparser` is set to `true`, omni will handle the
parsing of the arguments before calling the command, and put the results
of the parsing in environment variables readily available for the called
command.

The environment variables written are in the format:
- `OMNI_ARG_LIST` for the list of available arguments
- `OMNI_ARG_<argname>_TYPE` for the type of argument `argname`, having
  for values `int`, `float`, `bool` or `str`, and the variant
  `<type>/<size>` used when the argument `argname` is an array, in which
  cases values from 0 to `size` - 1 will be available.
- `OMNI_ARG_<argname>_VALUE` for the value of argument `argname`, if
  that argument is not an array, or
- `OMNI_ARG_<argname>_VALUE_<index>` for the value at index `index` of
  argument `argname` if that argument is an array

As to simplify the use of the argument parser in a native way for
languages, SDKs have been made available:
- https://github.com/omnicli/sdk-python
- https://github.com/omnicli/sdk-ruby
- https://github.com/omnicli/sdk-go

Closes: https://github.com/XaF/omni/issues/734