refactor(SyntaxOptArg): takes a single list of names
Previously, we had a separation between the primary name and aliases,
and we merged the information together before extracting the expected
main long name and main short name. This fixes that by merging
everything into a single 'names' list, and applying the same extraction
logic but avoiding the initial merging of the lists.

This also applies the splitting logic when defining a parameter from the
.omni.yaml file or the metadata yaml file, allowing to use the same
syntax as the `# opt` and `# arg` metadata headers.