refactor(`up`): ♻️ make environments versioned (#767)
When running `omni up`, the current environment of a repository will now
only be replaced upon success. This means that any failure to `omni up`
won't break the existing environment used in a repository (except for
potential side-effects). This also allows to keep an history of previous
environments for a configurable retention period, up to a configurable
maximum entries in the history per work directory, and globally.

Closes https://github.com/XaF/omni/issues/749