feat(`custom`): add support for adding environment variables
The `custom` operations of `omni up` now allow for adding environment
variables in the dynamic environment through the special `$OMNI_ENV`
environment variable provided during the `custom` operations.

This variable works similarly to `$GITHUB_ENV` in GitHub actions for
instance. It supports the basic environment setting format:
- `unset VAR` to unset a variable
- `VAR=value` to set a variable to the value `value`
- `VAR<<EOF` followed by any number of lines, followed by a delimiter
  (like `EOF`) to set a multiline environment variable using heredoc
  formatting (`VAR<<-EOF` also works to remove all indentation, and
  `VAR<<~EOF` to remove the minimum indentation)

It does also support extended operations. These are safer to use for the
dynamic environment provided by omni, as they are applied relative to
the environment at the time of applying the dynamic environment, vs.
being set as absolute values at the time of `omni up`.

The extended operations are the following:
- `VAR<<=value` to prepend a value to a path-like environment variable
- `VAR>>=value` to append a value to a path-like environment variable
- `VAR-=value` to remove a value from a path-like environment variable
- `VAR>=value` to add a suffix to an environment variable
- `VAR<=value` to add a prefix to an environment variable

Closes https://github.com/XaF/omni/issues/700