feat(`custom`): allow custom operations to add work directory binaries
With https://github.com/XaF/omni/pull/785 adding support for custom
environment variables when running a custom up operation, this is a
natural follow-up to allow custom operations to build any binary and
expose it in the dynamic environment built for the current work
directory.

This exposes the environment variables:
- `PREFIX` which represents the install prefix to be used
- `DESTDIR` which represents the destination directory where to
  put the binaries to be added to the custom environment

Anything added to `PREFIX` during the operation will simply be ignored.
And anything added to `DESTDIR` will be copied over to `PREFIX`. If a
custom operation has already been met, the existing contents of the
`PREFIX` directory will be kept around.

Closes https://github.com/XaF/omni/issues/701