feat(`custom`): allow custom operations to add work directory binaries (#788)
With https://github.com/XaF/omni/pull/785 adding support for custom
environment variables when running a custom up operation, this is a
natural follow-up to allow custom operations to build any binary and
expose it in the dynamic environment built for the current work
directory.

This exposes the environment variable `PREFIX` which represents the
install prefix to be used. This environment variable is available both
in the `met` and in the `meet` step, so it can be checked against to
decide if the `meet` step needs to be run.

If the `$PREFIX/bin` directory exists, it will be added to the `PATH`
automatically. If the `$PREFIX/lib` directory exists, it will be added
to the `LD_LIBRARY_PATH` automatically.

Closes https://github.com/XaF/omni/issues/701