chore(@-version): handle @-version as exact matching
The new `cargo-install` and `go-install` operations support specifying
versions using the `@` symbol, like when using the tool with either
`cargo install <crate>[@<version>]` or `go install <path>[@<version>]`,
but this was interpreted the same was as a weak version specification.

This changes that by considering that `@`-versions are an exact
specification.