fix(`cache`): 🐛 cleanup old versions for uninstalled resources
Both `github-release`, `go-install` and `cargo-install` keep caches of
the resolved versions for the given repositories, import paths and
crates respectively. These caches are useful when the resources are
actually in use, but if they are not in use for a while, and being
cleaned up by other rules, these versions should not stay around too
much.

This adds a retention period of 3 month per default for those versions,
after which they will be cleaned up during the cleanup step.

Fixes https://github.com/XaF/omni/issues/755