feat(`up`)!: 💥 replace `asdf` backend by `mise` (#832)
Until now, omni has been using `asdf` (https://asdf-vm.com/) as the
backend for supporting arbitrary tools that are not exposed directly by
GitHub releases, `go install` or `cargo install` calls. The latter two
also depend on `asdf`-installed `go` or `rust` to install the associated
tools.

Since the recent `asdf` change which moved the `asdf` install process to
depend on a go binary instead of a simple `git clone`, this would
require a refactor of the codebase to support installing/updating `asdf`
in its new form.

`mise` (https://mise.jdx.dev/) is a slightly more secure tool, written
in rust, which also supports natively pre-compiled versions of python
and handling rust using `rustup` directly. With omni requiring a backend
refactor in any case, might as well go for `mise` which also improves on
those points.

BREAKING CHANGE: Even with a number of automations to handle the
migration from `asdf` to `mise`, it is possible that at least one
instance of `omni up` in any repository is required, and maybe even
running `omni up` again in any repository where issues are encountered.
Also, it is possible that some plugins were working right of the bat
with `asdf` but might require specifying an `url:` when using with
`mise`; urls for base plugins can be found in
https://github.com/asdf-vm/asdf-plugins.

Closes https://github.com/XaF/omni/issues/830
Closes https://github.com/XaF/omni/issues/695