feat(`up`): ✨ add supply-chain checks (#848)
When using many sources to install tools for an environment, there can
be supply chain concerns. This aims at addressing those by adding
configuration to allow/deny specific operations for the `up` command, as
well as specific sources for those operations. This also allows to
allow/deny specific backends for mise, including a special "custom"
backend for user-provided or system-provided urls.

A special parameter also allows to set default URLs for specific tools,
allowing to override the plugin to be used by default for a tool.

Closes https://github.com/XaF/omni/issues/600