fix(`mise`): 🐛 add the `mise` binary in the `PATH` for `mise` commands
When `asdf` was used, we added the `asdf` binary in the `PATH` as some
plugins depend on being able to call `asdf`. For `mise`, we initially
removed that believing that the `mise` binary didn't depend on plugins
being able to call back `mise`, or that for some reason `mise` would add
itself to the `PATH` when operating on `asdf` plugins. This does not
seem to be the case however, and thus this adds the `mise/bin/`
directory to the `PATH`, which restore the behavior we had with `asdf`.