fix(`cargo-install` and `go-install`): `--no-cache` now works (#865)
The `--no-cache` option, when used with `cargo-install` and `go-install`
was failing with issues to replace binary files. This was caused by the
`bin` directory already existing.

On top of that, the `cargo-install` action was failing because of
replacement of the `rust` install backed by `mise`, since the right
environment variables were not put in place. This is fixed by adding
those environment variables.

Related to https://github.com/XaF/omni/issues/864