fix(`up`): initialize `omni_now()`
The `omni_now()` function is used to keep track of any data that is
fresh, i.e. that has been refreshed during the runtime of the `omni up`
command. That function initializes with the current time on the first
call and returns that time everytime after. To make sure that it is
accurate, we need to make sure it is called at least once before being
checked against.