feat(`up`): ✨ timeout and offer to kill running `up` if hanging (#879)
The `omni up` (and respectively `omni down`) command will try to attach
to any running `omni up` or `omni down` command that is already running
for the current work directory. However, in cases where there is a
connectivity issue (e.g. computer going to sleep), it is possible for
the `omni up` command to hang indefinitely (e.g. waiting for a network
response).

When attaching to a running process, this will trigger a timeout after a
few minutes if the process seems to be hanging, offering the user to
kill the running process.

Closes https://github.com/XaF/omni/issues/877