chore(`versioning`): move `omni` to use CalVer
With the evolution of `omni` and its stabilization, using SemVer does
not align as much as CalVer with the way of a CLI. More specifically,
omni is made to evolve forward, whether it is for security issues,
regular fixes or new features, we try to keep retrocompatibiliy of
evolution between versions, but we do not guarantee maintenability of
older versions in the long term.

CalVer thus represents a better versioning for omni. The versions of
omni will now follow the pattern `<year>.<short month>.<release>` where
`<release>` will be an incremental number reset to `0` each time the
month changes. We will not use any 0-padding in the version.