fix(`mise`): 🐛 replace `mise where` by `mise ls --installed`
This is used to figure out the normalized path to an installed tool as
returned by mise, but `mise where` seems to sometimes have weird
behaviors. Since we use it without caring for the version at all, using
`mise ls --installed` allows to return a path to any installed version
easily.