feat(`argparser`): ✨ add auto-completion support (#911)
When a command enables the `argparser`, omni will now provide
automatically `autocompletion` for that command using the information
that was provided to configure the `argparser`. This means that we can
auto-complete enum values for instance, as well as all parameters.

New argument types are introduced as `dir`, `file` and `repopath`
allowing to respectively auto-complete directories, files and repository
paths. When using those types, the value passed to the command will be
an absolute path.

Commands that have auto-completion enabled (`true`) will skip the
argparser auto-completion. There is also a new `partial` auto-completion
that is available when the argparser is configured, if the command
provider wants to offer extended auto-completion for the values of
specific parameters. The `partial` auto-completion will callout to the
underlying, trusted, command if and only if a _value_ should be
auto-completed for the current cursor position.

Closes https://github.com/xaf/omni/issues/854