feat(`packages`): ✨ make git updates more aggressive (#913)
By default, `git` updates for a repository are using fast-forward only
to avoid breaking anything that the user might have been working on.
This requirement does not apply to packages since packages are managed
by omni and omni only. We can thus be more aggressive when working in
packages by fetching any new commits, and resetting to the origin's
branch.

Closes https://github.com/xaf/omni/issues/908