fix(did you mean): `OMNI_SUBCOMMAND` set to selected command (#924)
At this time, the `OMNI_SUBCOMMAND` kept the original value used by the
caller before selecting the real command from the `Did you mean?` menu;
this led to issues for commands depending on `OMNI_SUBCOMMAND` to show
the help, for instance.

This is now fixed as it will set the `OMNI_SUBCOMMAND` environment
variable to the selected command in the `Did you mean?` menu.

Fixes https://github.com/xaf/omni/issues/919