feat(`config`): configuration commands can be exported in omnipath (#926)
By adding an omni configuration file in the omnipath, any command it
contains that have been flagged for exportation (`export: true`) will
now be available as part of the global commands, with the priority
corresponding to their position in the omnipath.

This will allow for easier writing of command wrappers for instance,
where a shell script wrapper could be replaced by a simple, exported
configuration command.

Closes https://github.com/xaf/omni/issues/925