feat(void command): support `-h` and `--help` for auto-generated commands (#929)
The `void commands` are auto-generated commands to be able to show the
help when there exist a sub-command, but no command was defined at this
level (e.g. `omni a b` is defined, but `omni a` isn't).

So far, calling `omni a --help` or `omni a -h` in that situation would
simply fail (and suggest, using `Did you mean?`, to call `omni a b`).
This changes that behavior by supporting the special `-h` and `--help`
arguments if these are the only ones being passed to a void command,
thus triggering to show the help message for that auto-generated command
in that situation.

Closes https://github.com/xaf/omni/issues/920