feat(`github release`): ✨ distinguish install configurations
When specifying configuration about which asset to install, we can
end-up with a release at a given version which is very different from
the release at the same version but with patterns to match assets.

This thus allow to distinguish installed github releases if any of the
configuration parameters that would impact the assets being installed is
used. This includes at this time disabling direct-binary-download,
skipping os or arch matching, preferring dist releases, or setting up
any pattern for the asset name.

The version as seen by the user will still be the same, but the install
directory will be suffixed by a `~` and a hash corresponding to the
configuration that was used to install that release.

Closes https://github.com/xaf/omni/issues/934