fix: 🐛 handle renames safely when not on same device (#948)
Currently all places `fs::rename` is called can fail if the paths are not
on the same device, as the documentation [0] states:
"This will not work if the new name is on a different mount point".

This fixes that by using `fs::copy` and `fs::remove` (and equivalent
to work on directories).

[0]: https://doc.rust-lang.org/std/fs/fn.rename.html