feat(python): ✨ use uv for venv and deps management
Until now, we were using `python -m venv` to handle the creation of virtual environments, and `pip install` for handling dependencies. We are changing this to use `uv` (https://github.com/astral-sh/uv) which is currently the reference in handling python environments as it builds virtual environments faster, and handles the management of dependencies faster too.

This also comes with another improvement as the `up`/`python` operation will now also be able to detect versions declared in a `pyproject.toml` file, as well as install dependencies from that file. On top of that, `pipenv` dependency files (`Pipfile` and `Pipfile.lock`) as well as `poetry` dependency files (`pyproject.toml` itself, but also `poetry.lock`) are now detected for installing dependencies automatically, making python environments ready by simply using `python: auto`.

Closes https://github.com/xaf/omni/issues/841
Closes https://github.com/xaf/omni/issues/295