feat!(`python`): ✨ default to install dependencies automatically
Previously, we were only installing dependencies automatically if the
python version requested was `auto`. With this change, we will default
to install dependencies from auto-detected well-known files no matter
which version of python was requested, as long as this is not
specifically disabled with `pip: false` or that files are not provided
with `pip: [file1, file2]`.