fix(autocomplete): 🐛 multi-word commands now autocomplete cleanly
When a multi-word command (e.g. "a b") existed, but it's "parent"
(e.g. "a") didn't, and if the multi-word command had
autocompletion
enabled through the argparser (or other means), the completion of
the full name of the command wouldn't happen, and the completion of
the arguments would kick-in too early. This fixes that.