feat(argparser): support list for parameter type
Add support for using a list directly as the parameter type to create
enum parameters, on top of existing syntaxes accepted for enum.

This provides:
  `type: [debug, info, warn, error]`
which is equivalent to the existing:
   `type: enum`
   with `values: [debug, info, warn, error]`
and to:
   `type: enum(debug, info, warn, error)`

If the `type` field is a list, the `values` field will be entirely
ignored, and the items in the `type` field will be used as `values`.