tests: split test files from the implementation (#1135)
The quantity of rust tests added in the codebase started to grow the
implementation files and make them harder to navigate. This uses a
pattern allowing to split the tests from the implementation, while
keeping them close. This is very similar to approaches taken in other
languages, for instance Go. It also allows to easily identify files that
don't have rust tests for now (and functionality that are thus only
covered by `bats`).