feat(gitignore): handle ignoring from workdir root
Omni already handles the global gitignore on top of the
local gitignore, but when a subdirectory of a workdir is
a git repository, the gitignore ends-up being checked from
the most specific git repository, and not necessarily the
one corresponding to the omni workdir.

This fixes that by passing down the workdir root path to be
used as git-base, if in a git workdir.

Closes https://github.com/xaf/omni/issues/1143