feat(`cd`): support full git URLs, and `--edit` flag to open the editor (#1229)
URLs in the shape `https://github.com/<org>/<repo>/tree/<branch>/<path>`
for instance contain all the information we need to identify the
repository, branch and path in there. Some URLs such as
`https://github.com/<org>/<repo>/blob/<branch>/<path>#L<lineno>` even
include a line number.

This adds support for `omni cd
https://github.com/<org>/<repo>/tree/<branch>/<path>` to get directly to
the right path, in the right repository. If the current branch is not
the requested one, omni will offer to switch branch before getting to
the directory.

This also adds support for `omni cd --edit
https://github.com/<org>/<repo>/blob/<branch>/<path>#L<lineno>` to open
the path, in the expected repository, in the default editor, if possible
at the provided line. Editors are searched first in the `VISUAL` then
`EDITOR` environment variables, then `vim` and `nano` are tried.

Closes https://github.com/xaf/omni/issues/401