feat(`cache`): clean stale environment versions after 6 months by default (#1235)
Omni was never cleaning the last used environment version for a work
directory that wasn't `omni down`. This could become a lot of entries
with the new `omni sandbox` as people could be playing around and simply
removing sandboxes without calling `omni down`.

This introduces a new clean-up process that will take care of removing
stale environment versions, i.e. environment versions that haven't been
`omni up`-ed in 6 months, on the condition that they are not found in
the file system. The duration is configurable using
`cache.environment.retention_stale`. The lookup on the file system
supports looking up in worktrees for git repositories, and in the
sandbox root.

This indirectly closes issue #78 which targeted the same concern, but in
a different way. The current approach is actually better as it does not
require users to run commands.

Closes https://github.com/xaf/omni/issues/78